Renamed OrthographyWord to OrthographyEditor
