+++ src/orthography/index.ts
@@ -1,3 +1,3 @@
 export * from './orthographyPopup';
 export * from './orthographyToggler';
-export * from './orthographyWord';
+export * from './orthographyEditor';
